Doors now hide items under them, unless they're open. If they're open they're displayed under objects.
For coders:
Layer = 2.4 (Pipes)
Layer = 2.5 (Wire)
Layer = 2.7 (open doors)
Layer = 2.8 (tables)
Layer = 2.9 (grille)
Layer = 3.0 (regular items and windows and most stuff, really)
Layer = 3.1 (closed, opening and closing door)
Layer = 4.0 (mobs)

There is a slight annomality that can occur but it shouldn't be too common, It happens when you have a closed fire door and a closed airlock over it. When you open the airlock, the opened airlock sets it's layer to 2.7, which is under the closed firedoor's layer 3.1, This causes the airlock to appear under the firelock while closed.

I also removed the window in the kitchen counter. Personally I think it looks better than the old, strange zig-zagging videow.

git-svn-id: http://tgstation13.googlecode.com/svn/trunk@1617 316c924e-a436-60f5-8080-3fe189b3f50e