Okay I'm done testing on the players, I've set the diseases for cold9 and spanish flu to their own disease path stuffs.
Grenades -
Quick patch because some one wanted to make an "allowed list", but rather than make the beakers loaded use the allowed list... they used a single item path (for single beakers).

It is just a few OR experssion so bottles and dispenser beakers can be used in grenade assemblies.

git-svn-id: http://tgstation13.googlecode.com/svn/trunk@1692 316c924e-a436-60f5-8080-3fe189b3f50e