When you're unconscious you will get the same blackness as always, but every 25 seconds, the alpha channel of it will reduce by about 30% for 0.3 seconds.
So... 25s blackness, 0.1s 15% transparent, 0.1s 30% transparent, 0.1s 15% transparent, repeat.

What this means is that you will be able to tell a tiny bit of what's going on around you and whether it's worth succumbing or not. It's a feature first introduced on Goon station and it's proven itself to be a good one.

git-svn-id: http://tgstation13.googlecode.com/svn/trunk@2360 316c924e-a436-60f5-8080-3fe189b3f50e