Command intercept will now only increase the security level if it's lower than blue...
Added a keycard authentication device. When you interact with it it asks you which event you'd like to trigger, currently only "Red alert". When you select the event, it asks you to swipe your ID card. If you have the new keycard auth access (60) on it (all heads of staff do), it will trigger all other devices of the same type to flash for 2 seconds. If someone with the same access level on their card swipes their card on another device during this period, the event will happen.

The idea behind this is the two key idea from submarine movies - just with ID cards. Admins are informed who requests the event and who triggers it btw, so abusing this is a bad idea.

git-svn-id: http://tgstation13.googlecode.com/svn/trunk@2635 316c924e-a436-60f5-8080-3fe189b3f50e