- commented out the old adamantine recipe. - removed adamantine ore - removed adamantine mineral walls - added a new adamantine recipe, but it's commented out for the moment - added a new material, mythril, with a nice urist sprite - added a mythril coin - added a mythril recipe, but it's commented out for the moment
So yeah, adamantine will have a new recipe and will no longer be gained from ore and added mythril to the procesisng unit, stacking unit and coin press.

NOTE TO URIST:
to enable this, you only have to uncomment the part with your name in machine_processing.dm Everything else is already in working order. Oh and you need to make the coin and sheet sprite.

git-svn-id: http://tgstation13.googlecode.com/svn/trunk@3146 316c924e-a436-60f5-8080-3fe189b3f50e