Changes to Antiqua.dmm
Widened area around engineering in case we will need to add anything there later. It looks ugly in the map editor, but nobody will notice in-game. We can turn it into a construction area for offices and whatnot in the meantime.

A few small changes to the medbay layout, attempted to give folks more room, and added more details and items.

Removed the old medbay area from the top side of the map.