TG: Added the caution code, sprited by magicbones, not present anywhere on the map
http://nanotrasen.com/phpBB3/viewtopic.php?f=10&t=4548&start=50#p37266

Added a power switch (/obj/structure/powerswitch), which is used to split or
connect two cables.

The way it works is simple. It connects the powernets by adding cable on the
floor it's on and removing it when it's turned off.

In-file description:
//This is a power switch. When turned on it looks at the cables around the tile
that it's on and notes which cables are trying to connect to it.
//After it knows this it creates the number of cables from the center to each of
the cables attempting to conenct. These cables cannot be removed
//with wirecutters. When the switch is turned off it removes all the cables on
the tile it's on.
//The switch uses a 5s delay to prevent powernet change spamming.

Screenshot:
http://www.kamletos.si/power%20switch.PNG
Revision: r3072
Author: 	 baloh.matevz