TG: (Paperwork overhaul, not actually added until we decide what to do with it.) Large bureaucracy overhaul. This affects clipboards, hand labelers, paper, paper bins, pens, photocopiers, and stamps. Everything (that I've changed significantly) should be BETTER.
NO paths have been changed.
All of the above have been added to the paperwork module (pending rename to
bureaucracy, which is COOLER).
SPRITES for most of the above (except photocopiers) are now in bureaucracy.dmi.

Clipboards in particular are MASSIVELY IMPROVED.
Hand labelers can now be turned ON AND OFF, to make it simpler to avoid
labelling tables.
This includes work in progress FOLDERS, which currently do NOTHING.
Revision: r3253
Author: 	 petethegoat