Captains no longer spawns with a cigar in their mouth, instead it spawns on the desk where the captain spawns.
CEs no longer spawns with a cigarette in their mouth. I did not add a cigarette to their desk because there is already a cigarette pack that spawns there.

Moved some stuff in singularity/New() into its own proc that singularity/New() calls. This is so that I can override it and prevent runtimes with an away mission.

git-svn-id: http://tgstation13.googlecode.com/svn/trunk@4717 316c924e-a436-60f5-8080-3fe189b3f50e