Removed the old gateway structure placeholders.
Made the gateway check for another gate before checking for whether it is charged. This is too stop poor confused players standing at the gate waiting for it to locate an away mission when one isn't loaded.

Ticked stationcollision.dm so it is actually possible to play Nodraks map without specifically compiling your server to do so.

git-svn-id: http://tgstation13.googlecode.com/svn/trunk@4870 316c924e-a436-60f5-8080-3fe189b3f50e