Added Exile Implants. Someone implanted with an Exile Implant will be able to enter the away mission, but unable to return from it. Not only can they be used for getting rid of dangerous criminals, but revs/stationheads count as dead while on the away mission, and traitor/changeling/wizard assassination targets count as dead if they're on the away mission at round end, allowing for those objectives to be completed peacefully.
Added a locker of exile implants to the gateway room.

Added a health analyzer to virology.

Replaced the teleport scroll in station collision with a cultist space suit/helmet. This is both because a teleport scroll is sorta overpowered in general and because it would allow people to easily return from exile.

git-svn-id: http://tgstation13.googlecode.com/svn/trunk@5133 316c924e-a436-60f5-8080-3fe189b3f50e