You can no longer steal chairs that mobs are buckled to those chairs. Fixes issue 1133.
False walls now have a variable that flips on when it's opening or closing. This should prevent the closing door code to execute after trying to open the door (or visa versa) due to the sleep()s used in there.

I've also re-organized teh  false_walls.dm file a bit. False r-walls really need to be changed into a subtype of false walls at some point.

git-svn-id: http://tgstation13.googlecode.com/svn/trunk@5289 316c924e-a436-60f5-8080-3fe189b3f50e