Re-added AI vocal announcements.
Unlike before, the announcement no longer require a huge list to work, saving up room from the string table.
They also work while streaming resources, meaning they won't overlap; this is because the resources are sent when a user joins, instead of when the AI makes an announcement and it messes up the ordering.

There have been improvements since last time, they now have a small delay of 10 seconds, which anyone is free to adjust.
The help screen will now only preview the sound to the AI.
The last announcement is saved to the AI's variables.
There's a limit of how many words can be said in a single announcement.